Allow lazy import of linker dispatchers
