+++ pytensor/link/numba/linker.py
@@ -1,12 +1,24 @@
+from functools import singledispatch
+
 from pytensor.link.basic import JITLinker
+
+
+@singledispatch
+def numba_typify(data, dtype=None, **kwargs):
+    raise NotImplementedError(
+        f"Numba funcify not implemented for data type {type(data)}"
+    )
+
+
+@singledispatch
+def numba_funcify(obj, *args, **kwargs):
+    raise NotImplementedError(f"Numba funcify not implemented for type {type(obj)}")
 
 
 class NumbaLinker(JITLinker):
     """A `Linker` that JIT-compiles NumPy-based operations using Numba."""
 
     def fgraph_convert(self, fgraph, **kwargs):
-        from pytensor.link.numba.dispatch import numba_funcify
-
         return numba_funcify(fgraph, **kwargs)
 
     def jit_compile(self, fn):
